+++ phpstan-bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 
 define('WPINC', 'tests/Stubs');
-define('GF_B_A_PLUGIN_FILE', 'plugin.php');
-define('GF_B_A_PLUGIN_SLUG', 'owc-gravityforms-bag-address');
-define('GF_B_A_ROOT_PATH', './');
-define('GF_B_A_VERSION', '1.0.0');
+define('GF_BAG_FILE', 'plugin.php');
+define('GF_BAG_SLUG', 'owc-gravityforms-bag-address');
+define('GF_BAG_ROOT_PATH', './');
+define('GF_BAG_VERSION', '1.0.1');