+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-### Changelog
+# Changelog
 
-requires: 4.9
-tested: 5.3.2
+All notable changes to this plugin will be documented in this file.
+
+* requires: 4.9
+* tested: 5.7.0
+
+## v1.0.1 - 2021-03-30
+
+### Removed
+- Province field
 
 ## v1.0.0